♻️ モジュール構成を変更
