+++ src/features/users/index.ts
@@ -3,8 +3,11 @@ export * from "./components/chef-follow-button";
 export * from "./components/chefs/chef-list/vertical-chef-list-item";
 export * from "./components/chefs/chef-list/vertical-chef-list";
 export * from "./components/chefs/chef-list/infinite-scroll-vertical-chef-list";
+export * from "./components/users/profile-form";
+
 //lib
 export * from "./lib/chefs/actions";
 export * from "./lib/users/actions";
 
 export * from "./types/chefs";
+export * from "./types/users";
